添加了有关Article的相关方法

+++ src/main/java/com/ssy/app/vo/ArticleVo.java
@@ -6,7 +6,7 @@ import io.swagger.annotations.ApiModelProperty;
 @ApiModel(value = "种草简介",description = "种草简介")
 public class ArticleVo {
     @ApiModelProperty(name = "id",value = "种草id")
-    private int id;
+    private long id;
     @ApiModelProperty(name = "nickName",value = "种草用户昵称")
     private String nickName;
     @ApiModelProperty(name = "timeAgo",value = "发布时间")
@@ -16,19 +16,19 @@ public class ArticleVo {
     @ApiModelProperty(name = "imgurl",value = "种草图片")
     private String imgurl;
     @ApiModelProperty(name = "starts",value = "收藏数")
-    private int starts;
+    private long starts;
     @ApiModelProperty(name = "likes",value = "点赞数")
-    private int likes;
+    private long likes;
     @ApiModelProperty(name = "comments",value = "评论数")
-    private int comments;
+    private long comments;
     @ApiModelProperty(name = "goodsInArticleVo",value = "商品信息")
     private GoodsInArticleVo goodsInArticleVo;
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
@@ -64,27 +64,27 @@ public class ArticleVo {
         this.imgurl = imgurl;
     }
 
-    public int getStarts() {
+    public long getStarts() {
         return starts;
     }
 
-    public void setStarts(int starts) {
+    public void setStarts(long starts) {
         this.starts = starts;
     }
 
-    public int getLikes() {
+    public long getLikes() {
         return likes;
     }
 
-    public void setLikes(int likes) {
+    public void setLikes(long likes) {
         this.likes = likes;
     }
 
-    public int getComments() {
+    public long getComments() {
         return comments;
     }
 
-    public void setComments(int comments) {
+    public void setComments(long comments) {
         this.comments = comments;
     }
